% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean_f.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
A number, which is the greatest common divisor.
}
\description{
Euclid's algorithm is an efficient method for computing the greatest common divisor of two numbers. 
The largest number that divides both of them without leaving a remainder.
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
